// microservices-kafka-elastisearch/order_service/src/dto/cartRequest.dto.ts
import { Static, Type } from "@sinclair/typebox";

export const CartRequestSchema = Type.Object({
	productId: Type.Integer(),
	customerId: Type.Integer(),
	qty: Type.Integer(),
});

export type CartRequestInput = Static<typeof CartRequestSchema>;

export const CartEditRequestSchema = Type.Object({
	id: Type.Integer(),
	qty: Type.Integer(),
});

export type CartEditRequestInput = Static<typeof CartEditRequestSchema>;